error handler

+++ src/main/java/com/decidir/sdk/dto/CustomerInSite.java
@@ -7,17 +7,17 @@ import java.io.Serializable;
  */
 public class CustomerInSite implements Serializable{
 
-    private Long days_in_site;
+    private Integer days_in_site;
     private Boolean is_guest;
     private String password;
-    private Long num_of_transactions;
+    private Integer num_of_transactions;
     private String cellphone_number;
 
-    public Long getDays_in_site() {
+    public Integer getDays_in_site() {
         return days_in_site;
     }
 
-    public void setDays_in_site(Long days_in_site) {
+    public void setDays_in_site(Integer days_in_site) {
         this.days_in_site = days_in_site;
     }
 
@@ -37,11 +37,11 @@ public class CustomerInSite implements Serializable{
         this.password = password;
     }
 
-    public Long getNum_of_transactions() {
+    public Integer getNum_of_transactions() {
         return num_of_transactions;
     }
 
-    public void setNum_of_transactions(Long num_of_transactions) {
+    public void setNum_of_transactions(Integer num_of_transactions) {
         this.num_of_transactions = num_of_transactions;
     }
